Use a constant time AES and GHASH implementation where needed
In the year of our Lord 2019, `crypto/aes` will use an AES
implementation that is essentially `rijndael-alg-fst.c` if the platform
does not have hardware support for the AES round function.

It is no longer the 20th century, so use a constant time AES and GHASH
on platforms where `crypto/aes` and GCM-AES are vartime.

`EnableVartimeAES()` will force the use of `crypto/aes` regardless of
security, for use cases where it does not matter.